move plugin into separate file
